learning/customer-invoice-api/requirements.txt
astroid==2.2.5
autoflake==1.3.1
blessings==1.7
bpython==0.18
certifi==2019.6.16
chardet==3.0.4
curtsies==0.3.0
Django==2.1
django-extensions==2.2.1
django-request-logging==0.7.0
djangorestframework==3.10.3
greenlet==0.4.15
httpie==1.0.3
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
pyflakes==2.1.1
Pygments==2.4.2
pylint==2.3.1
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
typing==3.7.4.1
urllib3==1.25.3
wcwidth==0.1.7
wrapt==1.11.2
